requirements.txt
matplotlib==3.2.1
numpy==1.18.5
torch==1.5.0
torchvision==0.5.0
pillow==7.1.2
gym==0.17.2
rdkit==2020.03.3
ase==3.19.1
xgboost==1.1.1
networkx==2.4